refactor: Use component-prefixed naming for module files
Rename module files to use consistent component-prefixed naming pattern:
- state.rs → *_state.rs (e.g., app_state.rs, search_state.rs)
- events.rs → *_events.rs (e.g., app_events.rs, editor_events.rs)
- content.rs → *_content.rs (e.g., help_content.rs, tooltip_content.rs)

This follows the same pattern established for *_render.rs files and makes
each file self-descriptive when viewing in editors or search results.
