+++ src/editor.rs
@@ -1,4 +1,4 @@
-pub mod events;
+pub mod editor_events;
 pub mod mode;
 
 pub use mode::EditorMode;
