+++ src/help.rs
@@ -2,10 +2,10 @@
 //!
 //! Contains the help popup state and content for keyboard shortcuts display.
 
-mod content;
+mod help_content;
 pub mod help_line_render;
 pub mod help_popup_render;
-mod state;
+mod help_state;
 
-pub use content::{HELP_ENTRIES, HELP_FOOTER};
-pub use state::HelpPopupState;
+pub use help_content::{HELP_ENTRIES, HELP_FOOTER};
+pub use help_state::HelpPopupState;
