+++ src/history.rs
@@ -1,7 +1,7 @@
-pub mod events;
+pub mod history_events;
 mod matcher;
 pub mod history_render;
-mod state;
+mod history_state;
 pub mod storage;
 
-pub use state::{HistoryState, MAX_VISIBLE_HISTORY};
+pub use history_state::{HistoryState, MAX_VISIBLE_HISTORY};
