+++ src/search.rs
@@ -3,9 +3,9 @@
 //! Provides text search functionality within the results pane.
 //! Users can search for text, see matches highlighted, and navigate between matches.
 
-pub mod events;
+pub mod search_events;
 mod matcher;
 pub mod search_render;
-mod state;
+mod search_state;
 
-pub use state::{Match, SearchState};
+pub use search_state::{Match, SearchState};
